requirements.txt
python-dotenv==1.1.1
groq==0.31.0
pefile==2024.8.26
requests==2.32.4
rich==13.7.1